tests: move util, co, validate-line-without-mode into lib
