+++ test/lib/util.js
