+++ test/lib/validate-line-without-mode.js
